added choice and fade out

+++ client/scripts/secret.js
@@ -22,23 +22,38 @@ $(document).ready(function (){
     });
 
 
-
     $("#writeData").on('click', 'button', function(){
-        $.ajax({
-            type:"DELETE",
-            url:"/things/" + $(this).data("id"),
-            success: function(){
 
-            },
-            error: function(xhr, status){
-                alert("Error: ", status );
-            },
-            complete: function(){
-                console.log("Delete Complete!");
 
-            }
+        if(confirm("Are you sure, you want to delete this post?") ){
+            $.ajax({
+                type: "DELETE",
+                url: "/things/" + $(this).data("id"),
+                success: function () {
+
+                },
+                error: function (xhr, status) {
+                    alert("Error: ", status);
+                },
+                complete: function () {
+                    console.log("Delete Complete!");
+
+                }
+            });
+
+        $(this).fadeTo("slow", 0.05, function(){ //fade
+
+            $(this).slideUp("slow", function() { //slide up
+
+                $(this).parent().remove();//then remove from the DOM
+
+            });
+
         });
-        $(this).parent().remove();
+
+        }
+        return false;
+
     });
     getData();
 });
@@ -56,8 +71,9 @@ function getData(){
 function appendToDom(data) {
     $('#writeData').empty();
     for (var i = 0; i < data.length; i++) {
-        $('#writeData').append("<div></div>")
+        $('#writeData').append("<div></div>");
         var $el = $('#writeData').children().last();
+
         $el.append("<p>A post from: " + data[i].name + "</p>");
         $el.append("<p> Your post: " + data[i].message + "</p>");
         $el.append("<p> Posted on: " + data[i].time + "</p>");
